Move new entity code to separate package
