+++ graylog2-server/src/main/java/org/graylog2/database/entities/EntityMetadata.java
@@ -14,7 +14,7 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-package org.graylog2.database.metadata;
+package org.graylog2.database.entities;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonCreator;
@@ -23,7 +23,6 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.google.auto.value.AutoValue;
 import org.joda.time.DateTime;
 
-import java.util.Collections;
 import java.util.Set;
 
 @AutoValue
